requirement.txt
torch>=1.6
torchvision
numpy
pandas
tqdm
lmdb
opencv-python
pillow
tensorboardx
wandb
scipy
tensorboardX>=1.14
clean-fid
blobfile
mpi4py